# api/openapi.yaml
openapi: 3.0.3
info:
  title: CargoFlows API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/me:
    get:
      operationId: getCurrentUser
      responses:
        '200': { description: Current authenticated user, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/change-password:
    post:
      operationId: changePassword
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChangePasswordRequest' } } }
      responses:
        '200': { description: Password changed and replacement bearer token issued, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /users:
    get:
      operationId: listUsers
      responses:
        '200': { description: Users, content: { application/json: { schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/User' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createUser
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateUserRequest' } } }
      responses:
        '201': { description: User created, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /users/{user_id}:
    parameters: [ { $ref: '#/components/parameters/UserID' } ]
    patch:
      operationId: updateUser
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateUserRequest' } } }
      responses:
        '200': { description: User updated, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      operationId: deleteUser
      description: Soft-delete a disabled admin or operator account while preserving historical attribution.
      responses:
        '204': { description: User deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{user_id}/password:
    parameters: [ { $ref: '#/components/parameters/UserID' } ]
    put:
      operationId: resetUserPassword
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResetUserPasswordRequest' } } }
      responses:
        '200': { description: Temporary password set, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /skus:
    get:
      operationId: listSKUs
      parameters:
        - { in: query, name: category_id, schema: { type: integer, minimum: 1 } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        '200': { description: SKUs, content: { application/json: { schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/SKU' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createSKU
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SKUMutationRequest' } } }
      responses:
        '201': { description: SKU created, content: { application/json: { schema: { $ref: '#/components/schemas/SKU' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /skus/{sku_id}:
    parameters: [ { $ref: '#/components/parameters/SKUID' } ]
    get:
      operationId: getSKU
      responses:
        '200': { description: SKU, content: { application/json: { schema: { $ref: '#/components/schemas/SKU' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      operationId: updateSKU
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SKUMutationRequest' } } }
      responses:
        '200': { description: SKU updated, content: { application/json: { schema: { $ref: '#/components/schemas/SKU' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      operationId: deleteSKU
      description: Deletes an unreferenced SKU. SKU records with inventory, media, model-family, AI, or published-content history must be disabled instead.
      responses:
        '204': { description: SKU deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: SKU is referenced by business history, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /skus/{sku_id}/inventory-adjustments:
    parameters: [ { $ref: '#/components/parameters/SKUID' } ]
    post:
      operationId: createInventoryAdjustment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateInventoryAdjustmentRequest' } } }
      responses:
        '201': { description: Inventory adjustment created, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryAdjustment' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /skus/{sku_id}/inventory-history:
    parameters: [ { $ref: '#/components/parameters/SKUID' } ]
    get:
      operationId: listInventoryHistory
      responses:
        '200': { description: Inventory history, content: { application/json: { schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/InventoryAdjustment' } } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /inventory-transactions:
    get:
      operationId: listInventoryTransactions
      parameters:
        - { in: query, name: type, schema: { type: string, enum: [purchase_receipt, sale_issue, customer_return, supplier_return, stock_adjustment] } }
        - { in: query, name: status, schema: { type: string, enum: [draft, posted] } }
        - { in: query, name: sku_id, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Inventory transactions, content: { application/json: { schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/InventoryTransaction' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createInventoryTransaction
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransactionRequest' } } } }
      responses:
        '201': { description: Draft created, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransaction' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /inventory-transactions/{transaction_id}:
    parameters: [ { in: path, name: transaction_id, required: true, schema: { type: string, format: uuid } } ]
    get:
      operationId: getInventoryTransaction
      responses:
        '200': { description: Inventory transaction, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransaction' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateInventoryTransaction
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransactionRequest' } } } }
      responses:
        '200': { description: Draft updated, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransaction' } } } }
  /inventory-transactions/{transaction_id}/post:
    parameters: [ { in: path, name: transaction_id, required: true, schema: { type: string, format: uuid } } ]
    post:
      operationId: postInventoryTransaction
      responses:
        '200': { description: Posted transaction, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransaction' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /inventory-transactions/{transaction_id}/reverse:
    parameters: [ { in: path, name: transaction_id, required: true, schema: { type: string, format: uuid } } ]
    post:
      operationId: reverseInventoryTransaction
      responses:
        '201': { description: Posted reversal, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryTransaction' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /inventory-transactions/{transaction_id}/corrections/preview:
    parameters: [ { in: path, name: transaction_id, required: true, schema: { type: string, format: uuid } } ]
    post:
      operationId: previewInventoryCorrection
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryCorrectionRequest' } } } }
      responses:
        '200': { description: Correction impact preview, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryCorrectionResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /inventory-transactions/{transaction_id}/corrections:
    parameters: [ { in: path, name: transaction_id, required: true, schema: { type: string, format: uuid } } ]
    get:
      operationId: listInventoryCorrections
      responses:
        '200': { description: Immutable correction versions, content: { application/json: { schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/InventoryCorrection' } } } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createInventoryCorrection
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryCorrectionRequest' } } } }
      responses:
        '201': { description: Correction created, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryCorrectionResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Idempotency or concurrent correction conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /brands:
    get:
      operationId: listBrands
      responses:
        '200': { description: Brands, content: { application/json: { schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/Brand' } } } } } } }
    post:
      operationId: createBrand
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 120 } } } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Brand' } } } } }
  /brands/{brand_id}:
    patch:
      operationId: updateBrand
      parameters: [{ in: path, name: brand_id, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 120 } } } } } }
      responses: { '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Brand' } } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    SKUID: { in: path, name: sku_id, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Unauthorized: { description: Missing or invalid bearer token, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorResponse' } } } }
...
    ValidationFailed: { description: SOP aggregate validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ValidationResponse' } } } }
...
  schemas:
    User:
      type: object
      additionalProperties: false
      required: [public_id, name, email, role, status, must_change_password, last_seen_at, created_at]
      properties:
        public_id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [super_admin, admin, operator] }
        status: { type: string, enum: [active, disabled] }
        must_change_password: { type: boolean }
        last_seen_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    AuthResponse:
      type: object
      additionalProperties: false
      required: [token, user]
      properties:
        token: { type: string, writeOnly: true }
        user: { $ref: '#/components/schemas/User' }
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password: { type: string, format: password, writeOnly: true }
        new_password: { type: string, format: password, minLength: 12, maxLength: 72, writeOnly: true }
...